# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'TIKETOK Web API'
  description: 'Web Resources Specification for TIKETOK'
servers:
  - url: http://localhost:8000/
    description: lbaw2464
...
paths:
...
  /events/{event_id}/join:
    post:
      operationId: joinEvent
      summary: 'Join an Event'
      description: 'Allows a user to join an event if it has available capacity.'
      tags:
        - 'Events'
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
            description: 'Unique identifier of the event.'
      responses:
        '200':
          description: 'Successfully joined the event.'
        '400':
          description: 'Event is at full capacity or user is already attending.'
  /events/{event_id}/leave:
    delete:
      operationId: leaveEvent
      summary: 'Leave an Event'
      description: 'Allows a user to leave an event they previously joined.'
      tags:
        - 'Events'
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
            description: 'Unique identifier of the event.'
      responses:
        '200':
          description: 'Successfully left the event.'
        '400':
          description: 'User is not part of this event.'

  # Comments
  /events/{event_id}/comments:
    post:
      operationId: addComment
      summary: 'Add Comment to Event'
      description: 'Allows users to add a comment to an event.'
      tags:
        - 'Comments'
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
            description: 'Unique identifier of the event.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: 'Comment added successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'

    delete:
      operationId: deleteComment
      summary: 'Delete Comment'
      description: 'Allows a user to delete their comment from an event.'
      tags:
        - 'Comments'
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
        - name: comment_id
          in: path
          required: true
          schema:
            type: integer
            description: 'Unique identifier of the comment.'
      responses:
        '204':
          description: 'Comment deleted successfully.'
components:
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        user_id:
          type: integer
        event_id:
          type: integer
    CommentInput:
      type: object
      properties:
        content:
          type: string